+++ src/yafowil/plone/autoform/events.py
@@ -1,12 +1,6 @@
-# -*- coding: utf-8 -*-
 from zope.interface import implementer
-try:
-    from zope.interface.interfaces import IObjectEvent
-    from zope.interface.interfaces import ObjectEvent
-except ImportError:
-    # BBB
-    from zope.component.interfaces import IObjectEvent
-    from zope.component.interfaces import ObjectEvent
+from zope.interface.interfaces import IObjectEvent
+from zope.interface.interfaces import ObjectEvent
 
 
 class IImmediateAddedEvent(IObjectEvent):