+++ src/yafowil/plone/connectors.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 from collections.abc import MutableMapping
 from zope.i18n import translate
 from zope.i18nmessageid import Message
@@ -8,11 +6,6 @@ from ZPublisher.HTTPRequest import HTTPRequest
 
 import six
 
-# Python 3.10
-try:
-    from collections import MutableMapping
-except ImportError:
-    from collections.abc import MutableMapping
 
 class Zope2RequestAdapter(MutableMapping):
     coding = "utf-8"